refactor(lapis): move some classes
Move some classes from `org.genspectrum.lapis.controller` to a more appropriate location.
I tried to leave only controllers and very closely related classes there.
