--- lapis/src/main/kotlin/org/genspectrum/lapis/controller/LapisResponse.kt
@@ -1,8 +1,0 @@
-package org.genspectrum.lapis.controller
-
-import org.genspectrum.lapis.request.LapisInfo
-import org.springframework.http.ProblemDetail
-
-data class LapisResponse<Data>(val data: Data, val info: LapisInfo = LapisInfo())
-
-data class LapisErrorResponse(val error: ProblemDetail, val info: LapisInfo = LapisInfo())
