+++ lapis/src/main/kotlin/org/genspectrum/lapis/openApi/OpenApiDocs.kt
@@ -16,35 +16,35 @@ import org.genspectrum.lapis.config.ReferenceGenomeSchema
 import org.genspectrum.lapis.config.SequenceFilterFieldName
 import org.genspectrum.lapis.config.SequenceFilterFieldType
 import org.genspectrum.lapis.config.SequenceFilterFields
-import org.genspectrum.lapis.controller.ACCESS_KEY_PROPERTY
 import org.genspectrum.lapis.controller.AGGREGATED_GROUP_BY_FIELDS_DESCRIPTION
-import org.genspectrum.lapis.controller.AMINO_ACID_INSERTIONS_PROPERTY
-import org.genspectrum.lapis.controller.AMINO_ACID_MUTATIONS_PROPERTY
 import org.genspectrum.lapis.controller.AMINO_ACID_MUTATION_DESCRIPTION
-import org.genspectrum.lapis.controller.COMPRESSION_PROPERTY
-import org.genspectrum.lapis.controller.Compression
 import org.genspectrum.lapis.controller.DETAILS_FIELDS_DESCRIPTION
-import org.genspectrum.lapis.controller.DOWNLOAD_AS_FILE_PROPERTY
-import org.genspectrum.lapis.controller.DataFormat
-import org.genspectrum.lapis.controller.FIELDS_PROPERTY
 import org.genspectrum.lapis.controller.FORMAT_DESCRIPTION
-import org.genspectrum.lapis.controller.FORMAT_PROPERTY
 import org.genspectrum.lapis.controller.LIMIT_DESCRIPTION
-import org.genspectrum.lapis.controller.LIMIT_PROPERTY
-import org.genspectrum.lapis.controller.MIN_PROPORTION_PROPERTY
-import org.genspectrum.lapis.controller.NUCLEOTIDE_INSERTIONS_PROPERTY
-import org.genspectrum.lapis.controller.NUCLEOTIDE_MUTATIONS_PROPERTY
 import org.genspectrum.lapis.controller.NUCLEOTIDE_MUTATION_DESCRIPTION
 import org.genspectrum.lapis.controller.OFFSET_DESCRIPTION
-import org.genspectrum.lapis.controller.OFFSET_PROPERTY
-import org.genspectrum.lapis.controller.ORDER_BY_PROPERTY
+import org.genspectrum.lapis.controller.middleware.Compression
+import org.genspectrum.lapis.controller.middleware.DataFormat
+import org.genspectrum.lapis.request.ACCESS_KEY_PROPERTY
+import org.genspectrum.lapis.request.AMINO_ACID_INSERTIONS_PROPERTY
+import org.genspectrum.lapis.request.AMINO_ACID_MUTATIONS_PROPERTY
 import org.genspectrum.lapis.request.AminoAcidInsertion
 import org.genspectrum.lapis.request.AminoAcidMutation
-import org.genspectrum.lapis.request.LapisInfo
+import org.genspectrum.lapis.request.COMPRESSION_PROPERTY
+import org.genspectrum.lapis.request.DOWNLOAD_AS_FILE_PROPERTY
+import org.genspectrum.lapis.request.FIELDS_PROPERTY
+import org.genspectrum.lapis.request.FORMAT_PROPERTY
+import org.genspectrum.lapis.request.LIMIT_PROPERTY
+import org.genspectrum.lapis.request.MIN_PROPORTION_PROPERTY
+import org.genspectrum.lapis.request.NUCLEOTIDE_INSERTIONS_PROPERTY
+import org.genspectrum.lapis.request.NUCLEOTIDE_MUTATIONS_PROPERTY
 import org.genspectrum.lapis.request.NucleotideInsertion
 import org.genspectrum.lapis.request.NucleotideMutation
+import org.genspectrum.lapis.request.OFFSET_PROPERTY
+import org.genspectrum.lapis.request.ORDER_BY_PROPERTY
 import org.genspectrum.lapis.request.OrderByField
 import org.genspectrum.lapis.response.COUNT_PROPERTY
+import org.genspectrum.lapis.response.LapisInfo
 import org.genspectrum.lapis.silo.ORDER_BY_RANDOM_FIELD_NAME
 
 fun buildOpenApiSchema(
